Type aliases.
This adds support for type aliases. Elm type aliases to existing types,
including generics, as well as aliases to tuples and units are generated
as Gleam aliases, while aliases to records and functions are defined as
types, because that's the only way they can be expressed in Gleam I'm
aware of.

<!-- ps-id: 542aaa75-a890-4547-be7a-187465c6e184 -->